backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv

# LangChain RAG
langchain
langchain-openai
langchain-community
unstructured[md]
chromadb
pypdf

# Supabase & Auth
supabase

# Testing
pytest
httpx
python-multipart
pytest-asyncio

# Formatting & Linting
black
ruff
